setup github codecov
