+++ package.json
@@ -4,15 +4,16 @@
   "description": "",
   "main": "src/index.js",
   "scripts": {
+    "test:jest":"jest ./src --coverage",
     "format": "prettier --write \"src/**/*.{js,jsx}\"",
     "lint": "eslint \"src/**/*.{js,jsx}\" --quiet",
     "testWatch": "mocha src/**/*.js --watch",
     "test": "mocha src/**/*.js",
-    "test:coverage":"npm run cover:unit && npm run cover:report",
-    "test:coveragehtml":"npm run cover:unit && npm run cover:reporthtml",
-    "cover:unit":"nyc --silent npm run test",
-    "cover:report":"nyc report --reporter=lcov --reporter=text --report-dir='./coverage'",
-    "cover:reporthtml":"nyc report --reporter-html --report-dir='./coverage'"
+    "test:coverage": "npm run cover:unit && npm run cover:report",
+    "test:coveragehtml": "npm run cover:unit && npm run cover:reporthtml",
+    "cover:unit": "nyc --silent npm run test",
+    "cover:report": "nyc report --reporter=lcov --reporter=text --report-dir='./coverage'",
+    "cover:reporthtml": "nyc report --reporter-html --report-dir='./coverage'"
   },
   "keywords": [],
   "author": "",
@@ -20,6 +21,7 @@
   "devDependencies": {
     "eslint": "^8.2.0",
     "eslint-config-prettier": "^8.3.0",
+    "jest": "^27.3.1",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
     "prettier": "^2.4.1"
